dom/media/mediasource/test/test_SplitAppend.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: append initialization and media segment separately</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  const sb = ms.addSourceBuffer("video/webm");

  const arrayBuffer = await fetchWithXHR("seek.webm");
  sb.appendBuffer(new Uint8Array(arrayBuffer, 0, 318));
  v.play();
  await once(sb, "updateend");
  sb.appendBuffer(new Uint8Array(arrayBuffer, 318));
  await once(sb, "updateend");
  ms.endOfStream();
  await once(v, "ended");
  // XXX: Duration should be exactly 4.0, see bug 1065207.
  ok(Math.abs(v.duration - 4) <= 0.002, "Video has correct duration");
  ok(Math.abs(v.currentTime - 4) <= 0.002, "Video has played to end");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
